requirements.txt
Flask==0.11.1
flask-cors==3.0.0
python-jose==1.2.0
requests==2.11.1